project_pipeline/requirements.txt
pytest==7.4.4
pandas==2.2.2
tqdm==4.66.4
requests==2.32.2
python-dotenv==0.21.0
nltk==3.9.1
emoji==2.14.0
snorkel
vaderSentiment==3.3.2

# For vertex AI
#torch==1.12.1
#transformers==4.21.0
#scikit-learn
google-cloud-storage
gcsfs
#accelerate==0.12.0
PyYAML>=6.0
tensorboard
kfp==2.4.0